Generic effects so that they can be applied on all components
